Remove JsonDB/JsonGuildDB/CoreDB and move parse_cli_flags
