+++ main.py
@@ -1,8 +1,7 @@
 from core.bot import Red, ExitCodes
 from core.global_checks import init_global_checks
 from core.events import init_events
-from core.settings import parse_cli_flags
-from core.cli import interactive_config, confirm
+from core.cli import interactive_config, confirm, parse_cli_flags
 from core.core_commands import Core
 from core.dev_commands import Dev
 import asyncio
